+++ lib/hooks/useRouterPath.ts
@@ -7,17 +7,3 @@ export function useRouterAsPath() {
     return router.asPath.split("?")[0] as string;
   }, [router.asPath]);
 }
-
-export function useRouterBasePath() {
-  const router = useRouter();
-  return useMemo(() => {
-    const path = router.asPath.split("/");
-
-    // For teams
-    if (path.length > 3) {
-      return `${path[1]}/${path[2]}`;
-    }
-
-    return path[1] as string;
-  }, [router.asPath]);
-}
